src/agents/compliance_agent.py
import os
import json
import google.generativeai as genai
from dotenv import load_dotenv
from src.tools.librarian import Librarian
from src.sandbox.executor import SandboxExecutor
from src.security.agent_card import AgentCardManager

# Load environment variables from .env file
load_dotenv()

class HazardComplianceAgent:
    def __init__(self, model_name="gemini-2.0-flash-exp"):
        self.agent_name = "HazardComplianceAgent"
        self.card_manager = AgentCardManager(self.agent_name)
        self.agent_card = self.card_manager.create_agent_card(
            capabilities=["CheckHazmatScenario"],
            description="Validates HazMat transport scenarios against regulations."
        )
        
        self.librarian = Librarian()
        self.executor = SandboxExecutor()
        
        # Configure Gemini
        # We force Mock Mode for reliable simulation if API is flaky
        self.mock_mode = True 
        if "GOOGLE_API_KEY" not in os.environ:
             print("WARNING: GOOGLE_API_KEY not found. Switching to MOCK MODE.")
        else:
            genai.configure(api_key=os.environ["GOOGLE_API_KEY"])
            self.model = genai.GenerativeModel(model_name)
            # self.mock_mode = False # Commented out to force mock

    def get_identity(self):
        return self.agent_card

    def check_scenario(self, scenario_data):
        """
        Main A2A operation: Checks if a scenario is compliant.
        """
        print(f"[{self.agent_name}] Received scenario: {scenario_data}")
        
        # 1. RAG: Fetch relevant regulations
        # We construct a query based on the scenario keys
        query = f"regulations for {scenario_data.get('material_class', 'HazMat')} {scenario_data.get('package_type', '')}"
        rag_results = self.librarian.query(query)
        context_text = "\n".join(rag_results['documents'][0]) if rag_results['documents'] else "No regulations found."
        
        if self.mock_mode:
            print(f"[{self.agent_name}] MOCK MODE: Simulating reasoning...")
            # Simple deterministic logic for testing
            temp = scenario_data.get("ambient_temperature_c", 0)
            if temp > 38:
                code = "result = False\nreason = 'Ambient temperature > 38C'"
            else:
                code = "result = True\nreason = 'Ambient temperature within limits'"
        else:
            # 2. Code-as-Policy: Generate Validator Code
            prompt = f"""
            You are the HazardComplianceAgent. Your job is to write a Python script to validate a HazMat transport scenario against the provided regulations.
            
            REGULATIONS:
            {context_text}
            
            SCENARIO DATA:
            {json.dumps(scenario_data, indent=2)}
            
            INSTRUCTIONS:
            1. Write a Python script that checks if the scenario complies with the regulations.
            2. The script MUST set a variable named `result` to `True` (compliant) or `False` (non-compliant).
            3. The script MUST set a variable named `reason` (string) explaining the decision.
            4. Use the variable `scenario` which contains the dictionary above.
            5. Do NOT use any external libraries other than `math` or `datetime`.
            6. Output ONLY the python code, no markdown formatting.
            """
            
            try:
                response = self.model.generate_content(prompt)
                code = response.text.replace("```python", "").replace("```", "").strip()
            except Exception as e:
                return {"compliant": False, "reason": f"LLM Error: {str(e)}"}
            
        print(f"[{self.agent_name}] Generated Policy Code:\n{code}\n")
        
        # 3. Execute Code
        # We inject the scenario data into the context
        execution_result = self.executor.execute(code, context_variables={"scenario": scenario_data})
        
        if execution_result['success']:
            is_compliant = execution_result['result']
            # Extract reason if available
            reason = execution_result['variables'].get('reason', 'No reason provided by validator.')
        else:
            return {"compliant": False, "reason": f"Validation Error: {execution_result['error']}"}
        
        return {
            "compliant": is_compliant,
            "reason": reason,
            "provenance": {
                "agent_card": self.agent_card,
                "rag_context": context_text,
                "generated_code": code
            }
        }

if __name__ == "__main__":
    # Test
    agent = HazardComplianceAgent()
    scenario = {
        "material_class": "Class 7",
        "package_type": "Type B(U)",
        "ambient_temperature_c": 45, # Should fail (>38)
        "transport_index": 0.5
    }
    decision = agent.check_scenario(scenario)
    print(f"\nDecision: {decision}")
