requirements.txt
google-generativeai
chromadb
pypdf
python-jose[cryptography]
python-dotenv
google-cloud-firestore
fpdf
langflow
